requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
future==0.18.2
gunicorn==20.0.4
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
torch==1.6.0+cpu
torchvision==0.7.0+cpu
whitenoise==5.2.0
